requirements.txt
appdirs==1.4.3
Django==1.11.1
django-extra-fields==0.9
djangorestframework==3.6.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
